Moving OzoneSecurityException to the exception package.
